feat(model): Send only changes when updating object, with PATCH
Update to accommodate angular-thickm 0.4.0-beta changes: when saving a model
with `save()`, send only the difference between the object gotten from server
and the one we currently have, and do it via PATCH. Most of the changes are in
angular-thickm, changes here are minor.

BREAKING CHANGE: Need to call `this.setOriginal()` in constructor, after doing
potential changes to the model. This is best done via the parent constructor,
i.e. `EveModel.call(this, data);` (which in turn uses the ThickModel
constructor).

Also removes `EveModel.prototype.fieldsToApi()`, which seemingly doesn't make
sense anymore -- it's now a part of `transformItemRequest()`.